polish > consolidated NewDocText into Text.js; Unused icons to be removed
