--- src/constants/arrows.js
@@ -1,5 +1,0 @@
-export const ARROW_SIZES = {
-  small: "small",
-  medium: "medium",
-  big: "big",
-};
